✨ feat: Phase 2 완료 - Backend 기반 구축 (T009-T020)
- PostEmbedding 엔티티 및 벡터 유사도 검색 구현
- UserRepository, PostEmbeddingRepository 생성
- PostRepository board-forum 스펙에 맞게 수정 (PostgreSQL 배열 연산자)
- Repository 영속성 테스트 13개 작성 (UserRepositoryTest 6개, PostRepositoryTest2 7개)
- SecurityConfig: Google OAuth2 + CORS 설정 완료
- JwtTokenService: JWT 토큰 생성/검증 (1시간/7일 유효)
- Phase 001 코드 백업 (.phase001.backup 확장자)

+++ backend/src/main/resources/db/migration/V5__create_posts_table.sql
@@ -6,11 +6,11 @@ DROP TABLE IF EXISTS post_embeddings CASCADE;
 DROP TABLE IF EXISTS posts CASCADE;
 
 CREATE TABLE posts (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     title VARCHAR(200) NOT NULL,
     content TEXT NOT NULL,               -- HTML 포맷 (Lexical 에디터 출력)
     plain_content TEXT NOT NULL,         -- Plain Text (벡터화 대상)
-    author_id INTEGER NOT NULL REFERENCES users(id) ON DELETE CASCADE,
+    author_id BIGINT NOT NULL REFERENCES users(id) ON DELETE CASCADE,
     hashtags TEXT[] DEFAULT '{}',
     created_at TIMESTAMP DEFAULT NOW(),
     updated_at TIMESTAMP DEFAULT NOW(),
